ENG-2600 refactor progress
